feat(plasma-core, plasma-ui, plasma-web): переносит общие компоненты в core
