+++ packages/plasma-core/src/components/VirtualCarousel/index.ts
@@ -1,10 +1,11 @@
-export { VirtualCarouselGridWrapper } from './VirtualCarousel';
 export { VirtualCarousel } from './VirtualCarousel';
-export { VirtualCarouselTrack } from './VirtualCarousel';
 
 export { VirtualCarouselContext } from './VirtualCarouselContext';
 
 export { VirtualCarouselItem } from './VirtualCarouselItem';
 export type { VirtualCarouselItemProps } from './VirtualCarouselItem';
+
+export { VirtualCarouselCol } from './VirtualCarouselCol';
+export type { VirtualCarouselColProps } from './VirtualCarouselCol';
 
 export type { VirtualCarouselProps } from './types';
