+++ packages/plasma-ui/src/components/VirtualCarousel/VirtualCarousel.stories.mdx
@@ -4,19 +4,20 @@ import { WithGridLines, InContainer } from '../../helpers/StoryDecorators';
 
 import * as stories from './VirtualCarousel.stories';
 
-import { VirtualCarouselGridWrapper, VirtualCarousel, VirtualCarouselItem, VirtualCarouselCol } from '.';
+import { CarouselGridWrapper } from '@salutejs/plasma-core';
+import { VirtualCarousel, VirtualCarouselItem, VirtualCarouselCol } from '.';
 
 <Meta title="Controls/VirtualCarousel" decorators={[InContainer, WithGridLines]} />
 
 # VirtualCarousel
-Набор компонентов для создания списков с прокруткой с виртуализацией.
+Набор компонентов для создания списков с прокруткой и виртуализацией.
 
 ## VirtualCarousel
 <Description of={VirtualCarousel} />
 <ArgsTable of={VirtualCarousel} exclude={["forwardedAs", "as"]} />
 
-## VirtualCarouselGridWrapper
-<Description of={VirtualCarouselGridWrapper} />
+## CarouselGridWrapper
+<Description of={CarouselGridWrapper} />
 
 ## VirtualCarouselItem
 <Description of={VirtualCarouselItem} />
