+++ packages/plasma-ui/src/components/VirtualCarousel/index.ts
@@ -1,13 +1,3 @@
-export { VirtualCarouselGridWrapper } from '@salutejs/plasma-core';
+export { VirtualCarouselItem, VirtualCarouselCol, VirtualCarousel, CarouselGridWrapper } from '@salutejs/plasma-core';
 
-export type { VirtualCarouselProps } from '@salutejs/plasma-core';
-
-export { VirtualCarouselItem } from '@salutejs/plasma-core';
-export type { VirtualCarouselItemProps } from '@salutejs/plasma-core';
-
-export { useRemoteHandlers2 } from './VirtualCarousel.hooks';
-
-export { VirtualCarousel } from './VirtualCarousel';
-
-export { VirtualCarouselCol } from './VirtualCarouselCol';
-export type { VirtualCarouselCarouselColProps } from './VirtualCarouselCol';
+export type { VirtualCarouselProps, VirtualCarouselItemProps, VirtualCarouselColProps } from '@salutejs/plasma-core';
